Allow using the service entity repository
