+++ src/Bundle/Doctrine/ORM/ResourceRepositoryTrait.php
@@ -1,0 +1,106 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Sylius\Bundle\ResourceBundle\Doctrine\ORM;
+
+use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\Mapping\ClassMetadata;
+use Doctrine\ORM\QueryBuilder;
+use Pagerfanta\Adapter\ArrayAdapter;
+use Pagerfanta\Doctrine\ORM\QueryAdapter;
+use Pagerfanta\Pagerfanta;
+use Sylius\Component\Resource\Model\ResourceInterface;
+
+/**
+ * @property EntityManagerInterface $_em
+ * @property ClassMetadata          $_class
+ *
+ * @method createQueryBuilder(string $alias, string $indexBy = null)
+ * @method find($id, $lockMode = null, $lockVersion = null)
+ */
+trait ResourceRepositoryTrait
+{
+    public function add(ResourceInterface $resource): void
+    {
+        $this->_em->persist($resource);
+        $this->_em->flush();
+    }
+
+    public function remove(ResourceInterface $resource): void
+    {
+        if (null !== $this->find($resource->getId())) {
+            $this->_em->remove($resource);
+            $this->_em->flush();
+        }
+    }
+
+    public function createPaginator(array $criteria = [], array $sorting = []): iterable
+    {
+        $queryBuilder = $this->createQueryBuilder('o');
+
+        $this->applyCriteria($queryBuilder, $criteria);
+        $this->applySorting($queryBuilder, $sorting);
+
+        return $this->getPaginator($queryBuilder);
+    }
+
+    protected function getPaginator(QueryBuilder $queryBuilder): Pagerfanta
+    {
+        // Use output walkers option in the query adapter should be false as it affects performance greatly (see sylius/sylius#3775)
+        return new Pagerfanta(new QueryAdapter($queryBuilder, false, false));
+    }
+
+    /**
+     * @param array $objects
+     */
+    protected function getArrayPaginator($objects): Pagerfanta
+    {
+        return new Pagerfanta(new ArrayAdapter($objects));
+    }
+
+    protected function applyCriteria(QueryBuilder $queryBuilder, array $criteria = []): void
+    {
+        foreach ($criteria as $property => $value) {
+            if (!in_array($property, array_merge($this->_class->getAssociationNames(), $this->_class->getFieldNames()), true)) {
+                continue;
+            }
+
+            $name = $this->getPropertyName($property);
+
+            if (null === $value) {
+                $queryBuilder->andWhere($queryBuilder->expr()->isNull($name));
+            } elseif (is_array($value)) {
+                $queryBuilder->andWhere($queryBuilder->expr()->in($name, $value));
+            } elseif ('' !== $value) {
+                $parameter = str_replace('.', '_', $property);
+                $queryBuilder
+                    ->andWhere($queryBuilder->expr()->eq($name, ':' . $parameter))
+                    ->setParameter($parameter, $value)
+                ;
+            }
+        }
+    }
+
+    protected function applySorting(QueryBuilder $queryBuilder, array $sorting = []): void
+    {
+        foreach ($sorting as $property => $order) {
+            if (!in_array($property, array_merge($this->_class->getAssociationNames(), $this->_class->getFieldNames()), true)) {
+                continue;
+            }
+
+            if (!empty($order)) {
+                $queryBuilder->addOrderBy($this->getPropertyName($property), $order);
+            }
+        }
+    }
+
+    protected function getPropertyName(string $name): string
+    {
+        if (false === strpos($name, '.')) {
+            return 'o' . '.' . $name;
+        }
+
+        return $name;
+    }
+}
