+++ src/Bundle/test/config/services.yaml
@@ -70,3 +70,5 @@ services:
             -
                 name: doctrine.event_subscriber
                 connection: default
+
+    App\Repository\ComicBookRepository: null
