+++ src/Bundle/test/src/Repository/ComicBookRepository.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Repository;
+
+use App\Entity\ComicBook;
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\Persistence\ManagerRegistry;
+use Sylius\Bundle\ResourceBundle\Doctrine\ORM\ResourceRepositoryTrait;
+use Sylius\Component\Resource\Repository\RepositoryInterface;
+
+final class ComicBookRepository extends ServiceEntityRepository implements RepositoryInterface
+{
+    use ResourceRepositoryTrait;
+
+    public function __construct(ManagerRegistry $registry)
+    {
+        parent::__construct($registry, ComicBook::class);
+    }
+}
